Joomla 4 fix installJoomlaMultilingualSite()
In Joomla 4 AssertionError: Timed out retrying after 4000ms: Expected to find element: #removeInstallationFolder, but never found it.

As there is no CSS id `removeInstallationFolder` existing in Joomla 4,
but there are two buttons (open frontend and open backend) with class
`complete-installation`are existing and clicking on one will complete
the installation and delete the folder.
Afterwards check `installation` is deleted by 404 in page title as working for Joomla 4 and 5.

Tested with Cypress 13.3.0, with Joomla 4.3.4 and Joomla 5.0.0-rc