installExtensionFromFolder() - obscure noscript
There is an obscure error from Cypress:
  assert expected <noscript> to be visible
  AssertionError
  Timed out retrying after 4000ms: expected '<noscript>' to be 'visible'
  This element <noscript> is not visible because it has an effective width and height of: 0 x 0 pixels.

No idea why Cypress is messing with the noscript tag. A working solution is to just cut out .should('be.visible').
However, I found a working solution and commented in detail, maybe someone understand this stuff and can solve cleanly.

Tested with Cypress 13.3.0, with Joomla 4.3.4 and Joomla 5.0.0-rc