publishModule() - needs click on actions first
Error from Cypress in support.js: checkForPhpNoticesOrWarnings()
  This element <joomla-toolbar-button#status-group-children-unpublish> is not visible because its parent <div.dropdown-menu.> has CSS property: display: none

It is necessary in publishModule() to click on the 'Actions' menu first to make the 'Unpublish' menu item visible.
And after 'Unpublish' we need to check all the results again.
The same applies to the next step before clicking 'Publish'.

Tested with Cypress 13.3.0, with Joomla 4.3.4 and Joomla 5.0.0-rc