checkForSystemMessage() added and unified
Since there are other places with the obscure "<noscript> be visible"
errors, as fixed in 1e6f451, this is solved in a new
checkForSystemMessage() support method.
And used everywhere where the System message is checked.
Along the way, a third workaround with additional access to the CSS class
was seen and is now used.