Minimised language selection for Joomla 6.0 (#47)
* Fix clickToolbarButton() test for Joomla 4.4

Checked was for CSS class `joomla-dialog-container, but in Joomla 4.4 the class name is `modal-title`.
To make the test running in Joomla 4.4 again, changed for checking for `<h3>` heading.

* [60] Minimised Language Selector

In Joomla 6.0, the minimised language selector was introduced (see https://github.com/joomla/joomla-cms/pull/44735).
This change ensures that the minimised language selector is handled correctly and works with previous Joomla versions as well.

* 1st jform_language needs to be deleted too